Fix. Db setting

+++ javascript/nodejs/nest-sls-sample-project/src/domain/member/member.controller.ts
@@ -1,9 +1,10 @@
 import {
+  BadRequestException,
   Body,
   Controller,
   Delete,
   Get,
-  HttpStatus,
+  HttpStatus, InternalServerErrorException,
   Param,
   Post,
   Put,
@@ -12,36 +13,31 @@ import { MemberJoinRequestDto } from './dto/member.join.request.dto';
 import { Member } from '../../entities/member/member';
 import { ApiOperation } from '@nestjs/swagger';
 import { MemberNosqlService } from './member.nosql.service';
+import { MemberSqlService } from './member.sql.service';
 
 const TAG = 'MEMBER_CONTROLLER';
 
 @Controller('/api/v1/individual/member')
 export class MemberController {
-  constructor(private readonly memberService: MemberNosqlService) {}
+  constructor(private readonly memberService: MemberSqlService) {}
 
   @ApiOperation({ summary: 'Register One Member Info' })
   @Post('/register/')
-  async registerMember(
-    @Body() data: MemberJoinRequestDto,
-  ): Promise<HttpStatus> {
-    if (!data) {
-      console.log(TAG, `REGISTER NEW MEMBER INFO IS EMPTY`);
-      return HttpStatus.BAD_REQUEST;
-    }
-    try {
-      const result: string = await this.memberService.create(data);
-      console.log(TAG, `Success Create Member: ${result}`);
-      return HttpStatus.OK;
-    } catch(e) {
-      console.log(TAG, e);
-      return HttpStatus.INTERNAL_SERVER_ERROR;
-    }
+  async registerMember(@Body() data: MemberJoinRequestDto): Promise<any> {
+    if (!data)
+      throw new BadRequestException(TAG, `Register New Member Info Is Empty`);
+
+    const result: string = await this.memberService.create(data);
+    console.log(TAG, `Success Create Member: ${result}`);
+    return HttpStatus.OK;
   }
 
   @ApiOperation({ summary: 'Get One Member Info By primary key' })
   @Get('/get/:pkey')
   async getMember(@Param('pkey') primaryKey: any): Promise<Member | null> {
-    console.log(TAG, `Get Member Info primaryKey : ${primaryKey}`);
+    if (!primaryKey)
+      throw new BadRequestException(TAG, `Get Member Info primaryKey: ${primaryKey}`);
+
     return this.memberService.get(primaryKey);
   }
 
@@ -54,28 +50,28 @@ export class MemberController {
 
   @ApiOperation({ summary: 'Get One Member Info By memberId' })
   @Get('/id/:memberId')
-  async getMemberById(
-    @Param('memberId') memberId: string,
-  ): Promise<Member | null> {
-    console.log(TAG, `Get Member Info memberId : ${memberId}`);
+  async getMemberById(@Param('memberId') memberId: string): Promise<Member | null> {
+    if (!memberId)
+      throw new BadRequestException(TAG, `Get Member Info memberId: ${memberId}`);
+
     return this.memberService.getById(memberId);
   }
 
   @ApiOperation({ summary: 'Get One Member Info By member email' })
   @Get('/email/:email')
-  async getMemberByEmail(
-    @Param('email') email: string,
-  ): Promise<Member[] | null> {
-    console.log(TAG, `Get Member Info memberId By email: ${email}`);
+  async getMemberByEmail(@Param('email') email: string): Promise<Member[] | null> {
+    if (!email)
+      throw new BadRequestException(TAG, `Get Member Info memberId By email: ${email}`);
+
     return this.memberService.getByEmail(email);
   }
 
   @ApiOperation({ summary: 'Get One Member Info By member nickname' })
   @Get('/nickName/:nickName')
-  async getMemberByNickName(
-    @Param('nickName') nickName: string,
-  ): Promise<Member[] | null> {
-    console.log(TAG, `Get Member Info memberId : ${nickName}`);
+  async getMemberByNickName(@Param('nickName') nickName: string): Promise<Member[] | null> {
+    if (!nickName)
+      throw new BadRequestException(TAG, `Get Member Info memberId : ${nickName}`);
+
     return this.memberService.getByNickName(nickName);
   }
 
@@ -92,7 +88,9 @@ export class MemberController {
     @Param('pkey') primaryKey: any,
     @Body() data: MemberJoinRequestDto,
   ): Promise<HttpStatus> {
-    console.log(TAG, `Update Members Info primaryKey : ${primaryKey}`);
+    if (!primaryKey)
+      throw new BadRequestException(TAG, `Update Members Info primaryKey : ${primaryKey}`);
+
     return (await this.memberService.update(data))
       ? HttpStatus.OK
       : HttpStatus.INTERNAL_SERVER_ERROR;
@@ -101,7 +99,9 @@ export class MemberController {
   @ApiOperation({ summary: 'Delete One Member Info By primary key' })
   @Delete('/delete/:pkey')
   async deleteMember(@Param('pkey') primaryKey: any): Promise<boolean> {
-    console.log(TAG, `Delete Member Info primaryKey : ${primaryKey}`);
+    if (!primaryKey)
+      throw new BadRequestException(TAG, `Delete Member Info primaryKey : ${primaryKey}`);
+
     return this.memberService.delete(primaryKey);
   }
 }
